% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdadmm.R
\name{calc_chi}
\alias{calc_chi}
\title{Calculate chi vector}
\usage{
calc_chi(beta, index, X, theta, sigma, y, z)
}
\arguments{
\item{beta}{current state of beta vector}

\item{index}{indicates which component of beta to update}

\item{X}{design matrix}

\item{theta}{current value of theta vector}

\item{sigma}{sigma constant}

\item{y}{y vector}

\item{z}{current value of z vector}
}
\value{
a n-vector for use in updating a component of beta
}
\description{
Calculate chi vector
}
