% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdadmm.R
\name{calc_chiiR}
\alias{calc_chiiR}
\title{Calculate chi vector}
\usage{
calc_chiiR(beta, j, Xrow, thetai, sigma, yi, zi)
}
\arguments{
\item{beta}{current state of beta vector}

\item{j}{indicates which component of beta to update}

\item{Xrow}{ith row of design matrix}

\item{thetai}{ith entry of current value of theta vector}

\item{sigma}{sigma constant}

\item{yi}{ith component of y vector}

\item{zi}{ith component of current value of z vector}
}
\value{
a 1-vector for use in updating a component of beta
}
\description{
Calculate chi vector
}
