% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmm.R
\name{check_criterion1R}
\alias{check_criterion1R}
\title{Check stopping criterion 1}
\usage{
check_criterion1R(X, beta, z, y, epsilon1, epsilon2)
}
\arguments{
\item{X}{design matrix}

\item{beta}{current value of beta vector}

\item{z}{current value of z vector}

\item{y}{y vector}

\item{epsilon1}{epsilon1 constant value}

\item{epsilon2}{epsilon2 constant value}
}
\value{
a logical vector of length one. TRUE if criterion 1 is satisfied and FALSE otherwise
}
\description{
Check stopping criterion 1
}
