% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmm.R
\name{check_criterion2}
\alias{check_criterion2}
\title{Check stopping criterion 2}
\usage{
check_criterion2(sigma, X, z, old_z, epsilon1, epsilon2, theta)
}
\arguments{
\item{sigma}{sigma constant}

\item{X}{design matrix}

\item{z}{current z vector (k)}

\item{old_z}{old z vector (k - 1)}

\item{epsilon1}{epsilon1 constant}

\item{epsilon2}{epsilon2 constant}

\item{theta}{current theta vector (k)}
}
\value{
a logical vector of length one. TRUE if criterion 2 is satisfied and FALSE otherwise
}
\description{
Check stopping criterion 2
}
