% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmm.R
\name{norm_vec}
\alias{norm_vec}
\title{Calculate L2 norm of a vector}
\usage{
norm_vec(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a numeric vector of length one, the l2 norm of x
}
\description{
Calculate L2 norm of a vector
}
\references{
https://stackoverflow.com/questions/10933945/how-to-calculate-the-euclidean-norm-of-a-vector-in-r
}
