% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{padmm}
\alias{padmm}
\title{Perform proximal ADMM for weighted L1-penalized quantile regression}
\usage{
padmm(
  beta0,
  z0,
  theta0,
  sigma,
  X,
  eta,
  y,
  l1,
  l2,
  w,
  nu,
  tau,
  gamma,
  max_iter,
  eps1,
  eps2
)
}
\arguments{
\item{beta0}{initial value of beta}

\item{z0}{initial value of z}

\item{theta0}{initial value of theta}

\item{sigma}{sigma constant, a positive number}

\item{X}{design matrix}

\item{eta}{eta constant}

\item{y}{y vector}

\item{l1}{L1 penalty parameter}

\item{l2}{L2 penalty parameter}

\item{w}{weights vector for lambda1 penalty}

\item{nu}{weights vector for lambda2 penalty}

\item{tau}{quantile, a number between 0 and 1}

\item{gamma}{gamma constant, affects the step length in the theta update step}

\item{eps1}{epsilon1 constant for stopping}

\item{eps2}{epsilon2 constant for stopping}

\item{maxiter}{maximum number of iterations}
}
\value{
beta, the vector of coefficient estimates
}
\description{
Perform proximal ADMM for weighted L1-penalized quantile regression
}
