% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{prox}
\alias{prox}
\title{Perform proximal mapping of rho tau}
\usage{
prox(xi, alpha, tau)
}
\arguments{
\item{xi}{a single number}

\item{alpha}{a number}

\item{tau}{a number between 0 and 1, the quantile of interest}
}
\value{
output of the proximal mapping of rho tau
}
\description{
Perform proximal mapping of rho tau
}
