% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdadmm.R
\name{qr_scdadmm_L1}
\alias{qr_scdadmm_L1}
\title{Perform scd ADMM for weighted L1-penalized quantile regression}
\usage{
qr_scdadmm_L1(
  beta0,
  z0,
  theta0,
  sigma = 1,
  X,
  eta = 1,
  y,
  lambda,
  w,
  tau,
  gamma = 0.1,
  maxiter = 10^5,
  epsilon1 = 0.001,
  epsilon2 = 0.001,
  epsilon3 = 0.001
)
}
\arguments{
\item{beta0}{initial value of beta}

\item{z0}{initial value of z}

\item{theta0}{initial value of theta}

\item{sigma}{sigma constant, a positive number}

\item{X}{design matrix}

\item{eta}{eta constant}

\item{y}{y vector}

\item{lambda}{L1 penalty constant}

\item{w}{weights vector}

\item{tau}{quantile, a number between 0 and 1}

\item{gamma}{gamma constant, affects the step length in the theta update step}

\item{maxiter}{maximum number of iterations}

\item{epsilon1}{epsilon1 constant for stopping}

\item{epsilon2}{epsilon2 constant for stopping}
}
\value{
beta, the vector of coefficient estimates
}
\description{
Perform scd ADMM for weighted L1-penalized quantile regression
}
