% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdadmm.R
\name{scdadmmR}
\alias{scdadmmR}
\title{Perform scd ADMM for weighted L1-penalized quantile regression}
\usage{
scdadmmR(
  beta0 = rep(1, ncol(X)),
  z0 = y - X \%*\% beta0,
  theta0 = rep(1, nrow(X)),
  sigma = 0.05,
  X,
  eta = eigen(t(X) \%*\% X)$values[1],
  y,
  lambda = 0.1,
  w = rep(1, length(beta0)),
  tau = 0.5,
  gamma = 1,
  max_iter = 10^5,
  epsilon1 = 0.001,
  epsilon2 = 0.001,
  epsilon3 = 0.001
)
}
\arguments{
\item{beta0}{initial value of beta}

\item{z0}{initial value of z}

\item{theta0}{initial value of theta}

\item{sigma}{sigma constant, a positive number}

\item{X}{design matrix}

\item{eta}{eta constant}

\item{y}{y vector}

\item{lambda}{L1 penalty constant}

\item{w}{weights vector}

\item{tau}{quantile, a number between 0 and 1}

\item{gamma}{gamma constant, affects the step length in the theta update step}

\item{epsilon1}{epsilon1 constant for stopping}

\item{epsilon2}{epsilon2 constant for stopping}

\item{epsilon3}{tolerance for scd updating of beta}

\item{maxiter}{maximum number of iterations}
}
\value{
beta, the vector of coefficient estimates
}
\description{
Perform scd ADMM for weighted L1-penalized quantile regression
}
