% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shrink}
\alias{shrink}
\title{Soft shrinkage operator per Gu et al. 2018}
\usage{
shrink(u, alpha)
}
\arguments{
\item{u}{a numeric vector of length one}

\item{alpha}{a numeric vector of length one}
}
\value{
a numeric vector of length one, the output of the soft shrinkage operator
}
\description{
Soft shrinkage operator per Gu et al. 2018
}
