% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrinkR}
\alias{shrinkR}
\title{Soft shrinkage operator}
\usage{
shrinkR(u, alpha)
}
\arguments{
\item{u}{a numeric vector of length one}

\item{alpha}{a numeric vector of length one}
}
\value{
a numeric vector of length one, the output of the soft shrinkage operator
}
\description{
Soft shrinkage operator
}
