% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_beta_padmm}
\alias{update_beta_padmm}
\title{Update beta for the proximal ADMM for weighted L1-penalized quantile regression}
\usage{
update_beta_padmm(beta, X, theta, sigma, eta, y, z, l1, l2, w, nu)
}
\arguments{
\item{beta}{current state of the beta parameter vector}

\item{X}{design matrix}

\item{theta}{current state of theta vector}

\item{sigma}{sigma constant, numeric vector of length 1}

\item{eta}{eta constant, numeric vector of length 1}

\item{y}{outcome vector}

\item{z}{current state of z vector}

\item{l1}{lambda1 penalty parameter}

\item{l2}{lambda2 penalty parameter}

\item{w}{weights vector for L1 penalty}

\item{nu}{weights vector for L2 penalty}
}
\value{
updated beta vector
}
\description{
Update beta for the proximal ADMM for weighted L1-penalized quantile regression
}
\seealso{
Other proximal ADMM for weighted L1 penalized quantile regression: 
\code{\link{update_zR}()},
\code{\link{update_z}()}
}
\concept{proximal ADMM for weighted L1 penalized quantile regression}
