% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmm.R
\name{update_beta_padmmR}
\alias{update_beta_padmmR}
\title{Update beta for the proximal ADMM for weighted elastic net-penalized quantile regression}
\usage{
update_beta_padmmR(
  beta,
  X,
  theta,
  sigma,
  eta,
  y,
  z,
  l1,
  l2 = 0,
  w,
  nu = rep(1, length(beta))
)
}
\arguments{
\item{beta}{current state of the beta parameter vector}

\item{X}{design matrix}

\item{theta}{current state of theta vector}

\item{sigma}{sigma constant, numeric vector of length 1}

\item{eta}{eta constant, numeric vector of length 1}

\item{y}{outcome vector}

\item{z}{current state of z vector}

\item{l1}{lambda1 penalty parameter}

\item{l2}{lambda2 penalty parameter}

\item{w}{weights vector with lambda1}

\item{nu}{weights vector with lambda2}
}
\value{
updated beta vector
}
\description{
Update beta for the proximal ADMM for weighted elastic net-penalized quantile regression
}
\concept{proximal ADMM for weighted elastic net penalized quantile regression}
