% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdadmm.R
\name{update_beta_scdadmmR}
\alias{update_beta_scdadmmR}
\title{One update of beta with scd}
\usage{
update_beta_scdadmmR(beta, X, theta, sigma, y, z, lambda, w)
}
\arguments{
\item{beta}{current value of beta vector}

\item{X}{design matrix}

\item{theta}{current value of theta vector}

\item{sigma}{sigma constant}

\item{y}{y vector}

\item{z}{current value of the z vector}

\item{lambda}{lambda constant}

\item{w}{w vector}
}
\value{
update of entirety of beta via scdadmm
}
\description{
One update of beta with scd
}
