% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmm.R
\name{update_theta}
\alias{update_theta}
\title{Update theta for the proximal ADMM or scd ADMM for weighted L1-penalized quantile regression}
\usage{
update_theta(theta, gamma, sigma, X, beta, z, y)
}
\arguments{
\item{theta}{current state of theta (k)}

\item{gamma}{gamma constant}

\item{sigma}{sigma constant}

\item{X}{design matrix}

\item{beta}{current state of beta (k + 1)}

\item{z}{current state of z (k + 1)}

\item{y}{y vector}
}
\value{
updated theta vector
}
\description{
Update theta for the proximal ADMM or scd ADMM for weighted L1-penalized quantile regression
}
