% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmm.R
\name{update_z}
\alias{update_z}
\title{Update z for the proximal ADMM or scd ADMM for weighted L1-penalized quantile regression}
\usage{
update_z(y, X, beta, theta, sigma, tau)
}
\arguments{
\item{y}{y vector}

\item{X}{design matrix}

\item{beta}{current state of beta parameter vector (k + 1)}

\item{theta}{current state of theta parameter vector (k)}

\item{sigma}{sigma constant}

\item{tau}{quantile, a number between 0 and 1}
}
\value{
updated z vector
}
\description{
Update z for the proximal ADMM or scd ADMM for weighted L1-penalized quantile regression
}
\seealso{
Other proximal ADMM for weighted L1 penalized quantile regression: 
\code{\link{update_beta_padmm}()}
}
\concept{proximal ADMM for weighted L1 penalized quantile regression}
