% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdadmm.R
\name{update_beta_component_scdadmm}
\alias{update_beta_component_scdadmm}
\title{Update component of beta with coordinate descent}
\usage{
update_beta_component_scdadmm(beta, index, X, theta, sigma, y, z, lambda, w)
}
\arguments{
\item{beta}{current value of beta vector, ie, a scalar}

\item{index}{an integer from 1 to p, indicating the index of beta for the update}

\item{X}{design matrix}

\item{theta}{current value of theta vector}

\item{sigma}{sigma constant}

\item{y}{y vector}

\item{z}{current value of the z vector}

\item{lambda}{lambda constant}

\item{w}{w vector}
}
\value{
update of one component of beta
}
\description{
Update component of beta with coordinate descent
}
