% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmm.R
\name{update_beta_padmmR}
\alias{update_beta_padmmR}
\title{Update beta for the proximal ADMM for weighted L1-penalized quantile regression}
\usage{
update_beta_padmmR(beta, X, theta, sigma, eta, y, z, lambda, w)
}
\arguments{
\item{beta}{current state of the beta parameter vector}

\item{X}{design matrix}

\item{theta}{current state of theta vector}

\item{sigma}{sigma constant, numeric vector of length 1}

\item{eta}{eta constant, numeric vector of length 1}

\item{y}{outcome vector}

\item{z}{current state of z vector}

\item{lambda}{lambda penalty parameter}

\item{w}{weights vector}
}
\value{
updated beta vector
}
\description{
Update beta for the proximal ADMM for weighted L1-penalized quantile regression
}
\seealso{
Other proximal ADMM for weighted L1 penalized quantile regression: 
\code{\link{update_zR}()}
}
\concept{proximal ADMM for weighted L1 penalized quantile regression}
